+++ src/cineripr/__init__.py
@@ -2,5 +2,4 @@ from __future__ import annotations
 
 __all__ = ["__version__"]
 
-__version__ = "2.5.17"
-
+__version__ = "2.5.19"